OT.O : config image version

+++ p3/scripts/install.sh
@@ -4,25 +4,36 @@
 
 set -e   # stop the script as soon as a command fails
 
-echo "==> 1/4  Installing curl"
-sudo apt-get update
-sudo apt-get install -y curl
+exist() {
+    command -v "$1" >/dev/null 2>&1
+}
 
-echo "==> 2/4  Installing Docker"
-curl -fsSL https://get.docker.com | sudo sh
-# Let my user talk to Docker without sudo (k3d needs this).
-sudo usermod -aG docker $USER
+if ! exist curl; then
+    echo "==> 1/4  Installing curl"
+    sudo apt-get update
+    sudo apt-get install -y curl
+fi
 
-echo "==> 3/4  Installing kubectl"
-KUBECTL_VERSION=$(curl -s https://dl.k8s.io/release/stable.txt)
-curl -Lo kubectl "https://dl.k8s.io/release/$KUBECTL_VERSION/bin/linux/amd64/kubectl"
-sudo install -m 755 kubectl /usr/local/bin/kubectl
-rm kubectl
+if ! exist docker; then
+    echo "==> 2/4  Installing Docker"
+    curl -fsSL https://get.docker.com | sudo sh
+    # Let my user talk to Docker without sudo (k3d needs this).
+    sudo usermod -aG docker $USER
+fi
 
-echo "==> 4/4  Installing k3d"
-curl -s https://raw.githubusercontent.com/k3d-io/k3d/main/install.sh | sudo bash
+if ! exist kubectl; then
+    echo "==> 3/4  Installing kubectl"
+    KUBECTL_VERSION=$(curl -s https://dl.k8s.io/release/stable.txt)
+    curl -Lo kubectl "https://dl.k8s.io/release/$KUBECTL_VERSION/bin/linux/amd64/kubectl"
+    sudo install -m 755 kubectl /usr/local/bin/kubectl
+    rm kubectl
+fi
 
-echo
+if ! exist k3d; then
+    echo "==> 4/4  Installing k3d"
+    choco install k3d --yes --no-progress
+fi
+
 echo "Installation finished:"
 docker --version
 kubectl version --client
